refactor: create addBook page
