--- ReactComponents/setupTest.js
@@ -1,3 +1,0 @@
-import matchers from '@testing-library/jest-dom/matchers';
-import { expect } from 'vitest';
-expect.extend(matchers);
