+++ ReactComponents/src/components/BookForm/BookForm.tsx
@@ -1,4 +1,3 @@
-import Collection from '../../pages/Collection/Collection';
 import React, { Component, createRef, RefObject } from 'react';
 import s from './BookForm.module.scss';
 import Categories from './Categories/Categories';
@@ -9,11 +8,11 @@ import SelectComponent from './SelectComponent/SelectComponent';
 import TextareaComponent from './TextareaComponent/TextareaComponent';
 import { InfoData } from 'app/types';
 
-interface State {
-  books: InfoData[];
+interface Props {
+  addBook: (value: InfoData) => void;
 }
 
-export default class BookForm extends Component<object, State> {
+export default class BookForm extends Component<Props> {
   private titleRef: RefObject<HTMLInputElement>;
   private isbnRef: RefObject<HTMLInputElement>;
   private pageCountRef: RefObject<HTMLInputElement>;
@@ -24,12 +23,8 @@ export default class BookForm extends Component<object, State> {
   private shortDescriptionRef: RefObject<HTMLTextAreaElement>;
   private longDescriptionRef: RefObject<HTMLTextAreaElement>;
   private downloadImgRef: RefObject<HTMLInputElement>;
-  constructor(props: object) {
+  constructor(props: Props) {
     super(props);
-
-    this.state = {
-      books: [],
-    };
 
     this.titleRef = createRef();
     this.isbnRef = createRef();
@@ -79,49 +74,46 @@ export default class BookForm extends Component<object, State> {
       status: this.statusRef.current?.value || '',
       categories: categoriesValues || [],
     };
-    this.setState({ books: this.state.books.concat(newBook) });
+    this.props.addBook(newBook);
     console.log(newBook);
     (e.target as HTMLFormElement).reset();
   };
 
   render() {
     return (
-      <>
-        <form onSubmit={this.handleAddBook} className={s.form}>
-          <InputText refLink={this.titleRef} name="title" label="Title" />
-          <InputText refLink={this.isbnRef} name="isbn" label="isbn" />
-          <InputAnother
-            refLink={this.pageCountRef}
-            name="pageCount"
-            label="pageCount"
-            type="number"
-          />
-          <InputText refLink={this.authorsRef} name="authors" label="authors" />
-          <TextareaComponent
-            refLink={this.shortDescriptionRef}
-            name="shortDescription"
-            rows={3}
-            label="shortDescription"
-          />
-          <TextareaComponent
-            refLink={this.longDescriptionRef}
-            name="longDescription"
-            rows={5}
-            label="longDescription"
-          />
-          <InputAnother
-            refLink={this.publishedDateRef}
-            name="publishedDate"
-            label="publishedDate"
-            type="date"
-          />
-          <DownloadImg refLink={this.downloadImgRef} />
-          <SelectComponent refLink={this.statusRef} name="status" />
-          <Categories refsLinks={this.checkboxRefs} name="categories" />
-          <button type="submit">add book</button>
-        </form>
-        <Collection books={this.state.books} search="" />
-      </>
+      <form onSubmit={this.handleAddBook} className={s.form}>
+        <InputText refLink={this.titleRef} name="title" label="Title" />
+        <InputText refLink={this.isbnRef} name="isbn" label="isbn" />
+        <InputAnother
+          refLink={this.pageCountRef}
+          name="pageCount"
+          label="pageCount"
+          type="number"
+        />
+        <InputText refLink={this.authorsRef} name="authors" label="authors" />
+        <TextareaComponent
+          refLink={this.shortDescriptionRef}
+          name="shortDescription"
+          rows={3}
+          label="shortDescription"
+        />
+        <TextareaComponent
+          refLink={this.longDescriptionRef}
+          name="longDescription"
+          rows={5}
+          label="longDescription"
+        />
+        <InputAnother
+          refLink={this.publishedDateRef}
+          name="publishedDate"
+          label="publishedDate"
+          type="date"
+        />
+        <DownloadImg refLink={this.downloadImgRef} />
+        <SelectComponent refLink={this.statusRef} name="status" />
+        <Categories refsLinks={this.checkboxRefs} name="categories" />
+        <button type="submit">add book</button>
+      </form>
     );
   }
 }
