+++ ReactComponents/src/pages/AddBook/AddBook.module.scss
@@ -1,0 +1,20 @@
+.form{
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding: 5px 0px;
+  width: 500px;
+  margin: 0 auto;
+  background: #2b2a2a;
+  padding: 15px;
+  font-size: 1rem;
+  border-radius: 3px;
+}
+.label{
+  //padding: 5px;
+  display: flex;
+  justify-content: space-between;
+  font-weight: bold;
+  border-bottom: 1px solid #8d8b8b;
+}
+
