+++ ReactComponents/src/pages/AddBook/AddBook.test.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import '@testing-library/jest-dom/extend-expect';
+import { render, screen } from '@testing-library/react';
+import Header from './BookForm';
+import { MemoryRouter } from 'react-router-dom';
+
+test('renders links to Home and About us', () => {
+  render(
+    <MemoryRouter>
+      <Header />
+    </MemoryRouter>
+  );
+
+  const homeLink = screen.getByRole('link', { name: 'Home' });
+  const aboutLink = screen.getByRole('link', { name: 'About us' });
+  expect(homeLink).toBeInTheDocument();
+  expect(aboutLink).toBeInTheDocument();
+});
