+++ ReactComponents/vite.config.js
@@ -3,21 +3,21 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 export default defineConfig({
-    plugins: [react()],
-    test: {
-        globals: true,
-        environment: 'jsdom',
-        setupFiles: './setupTest.ts',
-    },
-    // optimizeDeps: {
-    //   include: [
-    //     'src/**/*.ts',
-    //     'src/**/*.tsx',
-    //     'src/**/*.js',
-    //     'src/**/*.jsx',
-    //     'public/index.html',
-    //     'src/**/*.css',
-    //     'src/**/*.scss',
-    //   ],
-    // },
+  plugins: [react()],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: './setupTest.ts',
+  },
+  // optimizeDeps: {
+  //   include: [
+  //     'src/**/*.ts',
+  //     'src/**/*.tsx',
+  //     'src/**/*.js',
+  //     'src/**/*.jsx',
+  //     'public/index.html',
+  //     'src/**/*.css',
+  //     'src/**/*.scss',
+  //   ],
+  // },
 });
